requirements.txt
cryptography>=41.0.0
requests>=2.31.0
jsonschema>=4.19.0
click>=8.1.0
pydantic>=2.4.0
pytest>=7.0.0
aiohttp>=3.8.0
dnspython>=2.3.0
python-whois>=0.8.0 